requirements.txt
beautifulsoup4==4.10.0
pyppeteer==1.0.2
python-dotenv==0.19.2
pywhatkit==5.3
SQLAlchemy==1.4.31
